fix(types): review fixes — track-changes fallbacks, test consolidation, CI hardening
- Add || '' fallbacks for user.name/email/image in track-changes mark
  creation to prevent undefined author metadata when User fields are
  optional
- Consolidate overlapping consumer typecheck files: merge unique tests
  from index.ts into customer-scenario.ts, strip imports-main.ts to
  imports-only, delete redundant index.ts
- Pin PR preview checkout to exact SHA instead of branch ref
- Restore missing Editor import in EditorTypes.ts (lost during rebase)
