+++ tests/consumer-typecheck/src/imports-main.ts
@@ -101,72 +101,3 @@ import type {
   ProofingError,
   PageStyles,
 } from 'superdoc';
-
-// Verify types are usable (not just importable)
-const pageSize: PageSize = { w: 612, h: 792 };
-const margins: PageMargins = { top: 72, right: 72, bottom: 72, left: 72 };
-const saveOpts: SaveOptions = {
-  isFinalDoc: true,
-  fieldsHighlightColor: '#ff0',
-  compression: 'DEFLATE',
-};
-const binaryData: BinaryData = new ArrayBuffer(10);
-const rect: BoundingRect = {
-  top: 0,
-  left: 0,
-  bottom: 100,
-  right: 100,
-  width: 100,
-  height: 100,
-};
-const unsupported: UnsupportedContentItem = {
-  tagName: 'HR',
-  outerHTML: '<hr>',
-  count: 1,
-};
-
-// PresentationEditor event types
-const imgSelected: ImageSelectedEvent = {
-  element: document.createElement('div'),
-  blockId: 'block-1',
-  pmStart: 0,
-};
-const imgDeselected: ImageDeselectedEvent = { blockId: 'block-1' };
-const telemetry: TelemetryEvent = {
-  type: 'error',
-  data: { phase: 'render', error: new Error('test'), timestamp: Date.now() },
-};
-
-// Proofing types
-const proofingConfig: ProofingConfig = {
-  enabled: true,
-  defaultLanguage: 'en',
-  debounceMs: 500,
-};
-const proofingIssue: ProofingIssue = {
-  segmentId: 'seg-1',
-  start: 0,
-  end: 5,
-  kind: 'spelling' satisfies ProofingIssueKind,
-};
-const proofingStatus: ProofingStatus = 'idle';
-
-// FlowMode
-const flowMode: FlowMode = 'paginated';
-
-// Comment types (replaces customer's SuperComment / SuperElement overrides)
-const element: CommentElement = {
-  type: 'paragraph',
-  text: 'Hello',
-  content: [{ type: 'text', text: 'nested' }],
-};
-const comment: Comment = {
-  commentId: 'c-1',
-  createdTime: Date.now(),
-  creatorName: 'User',
-  creatorEmail: 'user@example.com',
-  elements: [element],
-  isDone: false,
-  parentCommentId: null,
-  importedId: 'imp-1',
-};
